Official: Exercises 1&2 Chapter 8
At 8.3 on the bottom are to be found such exercises.

(1) Given a list of integers, use a vector and return the median
(when sorted, the value in the middle position) and mode (the value
that occurs most often; a hash map will be helpful here) of the
list.

(2) Convert strings to pig latin. The first consonant of each word
is moved to the end of the word and “ay” is added, so “first”
becomes “irst-fay.” Words that start with a vowel have “hay” added
to the end instead (“apple” becomes “apple-hay”). Keep in mind the
details about UTF-8 encoding!