Official: Exercise 3 Ch8 + add some GOOD practice link
On the bottom of 8. 3 can be found this exercise.

(3) Using a hash map and vectors, create a text interface to allow a user to add
employee names to a department in a company. For example, “Add Sally to
Engineering” or “Add Amir to Sales.” Then let the user retrieve a list of all
people in a department or all people in the company by department, sorted
alphabetically.

The website 'exercism' has been added, it is clearly a very precious website for
a beginner like me. Besides a walkthrough with VERY KNORCKE projects has been
added as well.