Official: ch 12.3 - Start refactoring
The strategy to refactor the code is presented as a list of problems, then
problems 1,2 are addressed.

Precisely the responsibility is delegated away from the function main and the
configuration values are grouped together. We do it through these steps:

a. Extracting the arg parser
b. Grouping the config values into a Config struct