Oreilly: Chapter 10 - traits and generics
This chapter was read carefully but without trying harsh memorization or trying
the code with numerous examples. It was more important to have a vague idea of
the power of such feature of rust and how to use it, and then eventually
remember where to watch for answers in the chapter, in the case of a project in
progress.