Oreilly: skimming through Chapter 15
I write an average calculator for my strange academical situation
between two universities with two differetn grading systems and a
system of correction of votes which is not bijective. It was so
cool that I moved in a completely new repo on its own.

This exercise then gives the right push to study the chapters on
iterators from the oreilly book.

tmp - fixup rust project